requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.25
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
aiofiles==23.2.1
email-validator==2.1.0
bcrypt==4.0.1
msal==1.25.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
